requirements.txt
fastmcp>=1.1.0
chromadb>=0.4.24
langchain>=0.1.0
langchain-community>=0.0.10
openai>=1.12.0
pypdf>=3.17.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
nest-asyncio>=1.5.8